Move class ContainerEntry to a separate file
