+++ dotcms-integration/src/test/java/com/dotcms/rest/api/v1/page/PageResourceTest.java
@@ -49,7 +49,6 @@ import com.dotcms.rest.InitDataObject;
 import com.dotcms.rest.ResponseEntityView;
 import com.dotcms.rest.RestUtilTest;
 import com.dotcms.rest.WebResource;
-import com.dotcms.rest.api.v1.page.PageContainerForm.ContainerEntry;
 import com.dotcms.rest.api.v1.page.PageScenarioUtils.ContentConfig;
 import com.dotcms.rest.api.v1.personalization.PersonalizationPersonaPageView;
 import com.dotcms.util.FiltersUtil;
@@ -313,10 +312,9 @@ public class PageResourceTest {
         final ContentType bannerLikeContentType = TestDataUtils.getBannerLikeContentType();
         final Contentlet contentlet = TestDataUtils.getBannerLikeContent(true, 1, bannerLikeContentType.id(),
                 host);
-        final List<PageContainerForm.ContainerEntry> entries = new ArrayList<>();
+        final List<ContainerEntry> entries = new ArrayList<>();
         final String requestJson = null;
-        final PageContainerForm.ContainerEntry containerEntry =
-            new PageContainerForm.ContainerEntry(null, container.getIdentifier(), "1");
+        final ContainerEntry containerEntry = new ContainerEntry(null, container.getIdentifier(), "1");
         containerEntry.addContentId(contentlet.getIdentifier());
         entries.add(containerEntry);
         final PageContainerForm pageContainerForm = new PageContainerForm(entries, requestJson);
@@ -363,19 +361,18 @@ public class PageResourceTest {
             styleProperties.put("padding", "10px");
 
             // Create ContainerEntry with styleProperties
-            final List<PageContainerForm.ContainerEntry> entries = new ArrayList<>();
+            final List<ContainerEntry> entries = new ArrayList<>();
             final String containerUUID = UUIDGenerator.generateUuid();
             final Map<String, Map<String, Object>> stylePropertiesMap = new HashMap<>();
             stylePropertiesMap.put(contentlet.getIdentifier(), styleProperties);
 
-            final PageContainerForm.ContainerEntry containerEntry =
-                    new PageContainerForm.ContainerEntry(
-                            null,
-                            container.getIdentifier(),
-                            containerUUID,
-                            list(contentlet.getIdentifier()),
-                            stylePropertiesMap
-                    );
+            final ContainerEntry containerEntry = new ContainerEntry(
+                    null,
+                    container.getIdentifier(),
+                    containerUUID,
+                    list(contentlet.getIdentifier()),
+                    stylePropertiesMap
+            );
 
             entries.add(containerEntry);
             final PageContainerForm pageContainerForm = new PageContainerForm(entries, null);
@@ -489,7 +486,7 @@ public class PageResourceTest {
         final PageContainerForm form = mapper.readValue(json, PageContainerForm.class);
 
         assertNotNull("Form should not be null", form);
-        final PageContainerForm.ContainerEntry containerEntry = form.getContainerEntries().get(0);
+        final ContainerEntry containerEntry = form.getContainerEntries().get(0);
         assertNotNull("StylePropertiesMap should not be null", containerEntry.getStylePropertiesMap());
         assertTrue("StylePropertiesMap should be empty", containerEntry.getStylePropertiesMap().isEmpty());
     }
@@ -1457,9 +1454,8 @@ public class PageResourceTest {
      */
     private PageContainerForm createPageContainerForm(final String containerId, final List<String> contentletIds,
                                                       final String containerUUID) {
-        final List<PageContainerForm.ContainerEntry> entries = new ArrayList<>();
-        final PageContainerForm.ContainerEntry containerEntry = new PageContainerForm.ContainerEntry(null,
-                containerId, containerUUID);
+        final List<ContainerEntry> entries = new ArrayList<>();
+        final ContainerEntry containerEntry = new ContainerEntry(null, containerId, containerUUID);
         contentletIds.forEach(containerEntry::addContentId);
         entries.add(containerEntry);
         return new PageContainerForm(entries, null);
